Added optional input parameters to customize the package structure
Currently all the interfaces are getting created under resources folder
and similarly the pojo's are getting created under model folder. I have
modified the code to take these folder name as  optional input
parameters ,so that if provided the code get created as per the inputs
,if these optional parameters are not provided then the tool will behave
as it behaves now.